payment_service/requirements.txt
Flask==2.1.0
Flask-PyMongo==2.3.0
Werkzeug==2.1.0
PyJWT==2.3.0
pika==1.2.1